Split home/availability.py into availability/formatting.py, availability/overlap.py, and home/team_availability.py
Separates pure slot-formatting helpers, generic slot/overlap math, and
team-formation-specific overlap logic (which depends on home.models) into
distinct modules so the availability app doesn't depend on home.
