+++ availability/tests/test_overlap.py
@@ -4,13 +4,12 @@ from django.test import TestCase
 
 from accounts.factories import UserFactory
 from availability.factories import UserAvailabilityFactory
-from home.availability import (
+from availability.overlap import (
     AvailabilityWindow,
     calculate_overlap,
-    calculate_team_overlap,
+    calculate_user_overlap,
     count_one_hour_blocks,
-    format_slot_as_time,
-    format_slots_as_ranges,
+    find_best_one_hour_windows,
 )
 
 
@@ -59,6 +58,10 @@ class AvailabilityUtilsTestCase(TestCase):
         # Empty slots
         self.assertEqual(count_one_hour_blocks([]), 0)
 
+        # Gap before a consecutive pair
+        slots = [10.0, 11.0, 11.5]
+        self.assertEqual(count_one_hour_blocks(slots), 1)
+
     def test_calculate_overlap(self):
         """Test overlap calculation for groups and pairs."""
         # User1 and User2 overlap on Monday 12:00-15:00 (6 slots = 3 hours)
@@ -80,76 +83,46 @@ class AvailabilityUtilsTestCase(TestCase):
         slots, hours = calculate_overlap([self.user1, self.user3])
         self.assertEqual(hours, 0)  # No overlap because user3 has no availability
 
-    def test_calculate_team_overlap(self):
-        """Test team overlap calculation."""
-        # Create a captain with different availability
-        captain = UserFactory(username="captain", email="captain@example.com")
-        # Captain: Monday 11:00-14:00 (6 slots = 3 hours)
-        UserAvailabilityFactory(
-            user=captain, slots=[35.0 + (i * 0.5) for i in range(6)]
+        # No users
+        self.assertEqual(calculate_overlap([]), ([], 0))
+
+    def test_calculate_user_overlap(self):
+        """Test the two-user convenience wrapper around calculate_overlap."""
+        slots = calculate_user_overlap(self.user1, self.user2)
+        self.assertEqual(len(slots), 6)
+
+    def test_find_best_one_hour_windows(self):
+        """Test ranking one-hour windows by how many users are available."""
+        windows = find_best_one_hour_windows([self.user1, self.user2], top_n=2)
+        self.assertEqual(len(windows), 2)
+        self.assertTrue(all(window.total_available == 2 for window in windows))
+        self.assertIn("Mon", windows[0].formatted_time)
+
+    def test_availability_window_properties(self):
+        """Test the AvailabilityWindow dataclass' derived properties."""
+        window = AvailabilityWindow(
+            slot_range=(34.0, 34.5),  # Monday 10:00-11:00
+            formatted_time="Mon 10:00 AM - 11:00 AM",
+            available_users=[self.user1, self.user2],
+            unavailable_users=[self.user3],
+            role_counts={"Captain": 1, "Navigator": 0},
         )
-
-        result = calculate_team_overlap(
-            navigator_users=[self.user1],
-            captain_user=captain,
-            djangonaut_users=[self.user2],
-        )
-
-        # Navigator + djangonauts overlap (user1 + user2)
-        self.assertEqual(result["navigator_meeting_hours"], 3)
-        # user1 and user2 overlap is 3 hours, which is < 5 hours required
-        self.assertFalse(result["is_valid"])
-
-        # Check captain meetings
-        self.assertEqual(len(result["captain_meetings"]), 1)
-        # Captain 1-on-1 with user2
-
-    def test_format_slot_as_time(self):
-        """Test time formatting."""
-        # Sunday 00:00 (12:00 AM)
-        self.assertEqual(format_slot_as_time(0.0), "Sun 12:00 AM")
-
-        # Monday 14:30 (2:30 PM)
-        self.assertEqual(format_slot_as_time(38.5), "Mon 2:30 PM")
-
-        # Saturday 23:30 (11:30 PM)
-        self.assertEqual(format_slot_as_time(167.5), "Sat 11:30 PM")
-
-    def test_format_slots_as_ranges(self):
-        """Test formatting slots as time ranges."""
-        # Consecutive slots
-        slots = [10.0, 10.5, 11.0, 11.5]
-        ranges = format_slots_as_ranges(slots)
-        self.assertEqual(len(ranges), 1)
-        self.assertIn("Sun", ranges[0])
-
-        # Non-consecutive slots
-        slots = [10.0, 10.5, 12.0, 12.5]
-        ranges = format_slots_as_ranges(slots)
-        self.assertEqual(len(ranges), 2)
-
-        # Empty slots
-        ranges = format_slots_as_ranges([])
-        self.assertEqual(ranges, [])
-
-
-class AvailabilityWindowTestCase(TestCase):
-    """Tests for AvailabilityWindow dataclass."""
+        self.assertEqual(window.total_available, 2)
+        self.assertEqual(window.role_summary, "Captain: 1")
+        self.assertEqual(window.unavailable_member_ids, [self.user3.id])
+        self.assertEqual(window.start_datetime.weekday(), 0)  # Monday
 
     def test_admin_unavailable_url_uses_user_ids(self):
         """Test that admin_unavailable_url uses user.id, not str(user)."""
-        user1 = UserFactory(username="user1", email="user1@example.com")
-        user2 = UserFactory(username="user2", email="user2@example.com")
-
         window = AvailabilityWindow(
             slot_range=(10.0, 10.5),
             formatted_time="Sun 10:00 AM - 11:00 AM",
             available_users=[],
-            unavailable_users=[user1, user2],
+            unavailable_users=[self.user1, self.user2],
         )
 
         url = window.admin_unavailable_url
-        expected_ids = f"{user1.id},{user2.id}"
+        expected_ids = f"{self.user1.id},{self.user2.id}"
         self.assertIn(f"?user_id__in={expected_ids}", url)
         self.assertIn("home/sessionmembership/", url)
